Removed 'filename' member from MappedDirEntry struct:
MapperDirEntry stored both 'filename' and 'shortname'. 'filename' is always
equal to 'hostDir/shortname', so it's not needed to store it. If you don't
store it you also don't need to keep it in sync. In fact I found one place in
the code were we forgot to do this.

Unfortunately both 'filename' and 'shortname' are also stored in the
sector-cache. I kept it like this, but maybe we should still change this before
the next release? (Or remove it completely?)


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@7654 7512d703-071c-0410-b5c2-d2bdd895b7d1