made Connector::getPluggable() non-virtual
The only reason why this was virtual is that it allows to use covariant return
types. Something similar can be done by creating a specific getter method in
the derived classes (with a different name, but this may even be clearer).

Not using covariant return types also allowed to reduce #include dependencies.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@7686 7512d703-071c-0410-b5c2-d2bdd895b7d1