made deserialize more similar to regular machine load:
Before we used serialization framework to create and initialize the
devices. Now we use the regular createDevice() method to create the
devices and the serialization only to initialize the devices. This has
the advantage that all intermediate steps are consistent with the
regular load (e.g. the availableDevices vector in MSXMotherBoard).

This was needed for accurate device-references (used by ADVram). It will
also be needed for correct error handling, for example when loading a
savestate fails because some ROM is now missing.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8002 7512d703-071c-0410-b5c2-d2bdd895b7d1