made SectorBasedDisk class stateless wrt raw track writing:
IOW moved the format-in-progress-state from the disk image class to the
controller class (is also done like this for regular sector writes).

This fixes a problem with savestates in the middle of a disk format,
because we don't serialize the whole disk image classes (instead we only
store the filename of the diskimage and reinsert that disk on load). The
state of the controller class is serialized.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8080 7512d703-071c-0410-b5c2-d2bdd895b7d1