fix flush in debugdevice:
Implementation didn't follow the documented behaviour in multi-byte
ascii mode (flush on mode register write).
Though instead of fixing it like this, we simply flush after every
byte now. This makes the device easier to understand and use (I've
also updated the documentation to match this new behaviour).

Thanks to yakumoklesk for reporting this bug.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8180 7512d703-071c-0410-b5c2-d2bdd895b7d1