[6/7] cleanups: use the c++11 'auto' feature more
This is a relatively large patch, but it's also relatively simple. It replaces
stuff like:
   ReverseManager& reverseManager = motherBoard.getReverseManager();
with:
   auto& reverseManager = motherBoard.getReverseManager();
so we let the compiler automatically deduce the type of the 'reverseManager'
variable. Both code fragments mean exactly the same, the only difference is
that the programmer doesn't explicitly mention the type.

  We already did use this feature in our code, this patch only uses it a bit
  more. There are still lots more places where it could be used to simplify the
  code. I didn't try to find all those places. What I did in this patch was
  extract all these changes from my private c++11 branch and (only) in the
  already changed files look for more opportunities like this.

Some people argue that using this feature makes the code harder to understand
because you don't explicitly see the type of the variables anymore. For example:
  FooBar& bar = qux.getBar();
  bar.foo();
vs
  auto& bar = qux.getBar();
  bar.foo();
In the second version you have to lookup the declaration of getBar() to figure
out that the type of 'bar' is 'FooBar'. Popular counter arguments are that you
should pick meaningful variable and method names so that the type is obvious
(like in the example at the top). Or that it's not always required to know the
exact type. For example few people complain that the following construct is
hard to understand:
  qux.getBar().foo();
While in fact it's basically the same as above, the only difference is that
above we gave the first subexpression a name.

Anyway, like any language feature, 'auto' can be used to write better code or
abused to make the code worse. Since this is a new feature we still have to
gain experience in using it well. In this patch I tried to only use it when it
makes the code 'better' (and of course good vs bad is not well defined in this
context).

--

In this patch I did partly change some of the uses of 'auto' we introduced
earlier. For example:
  auto device = motherBoard.findDevice(name);
vs
  auto* device = motherBoard.findDevice(name);
Again both fragments mean exactly the same, in both cases the type of 'device'
is deduced to be a MSXDevice-pointer. Personally I find it useful to explicitly
state whether the type is a pointer-, reference- or value-type. For example
because (by convention) pointers can be nullptr but references cannot (so this
info is useful on code reviews to verify if there's a potential missing nullptr
check).

--

Using auto to write type-agnostic code can also make the code easier to
maintain or remove (small) hidden performance issues. For example
  std::string getName();        // original
  ...
  std::string name = getName();
  if (name.empty()) { ...
Maybe later code refactorings change the function getName() to
  const std::string& getName(); // 1st alternative
or maybe
  string_ref getName();         // 2nd alternative
In the 1st alternative, the caller code will continue to compile, but it does
make an unnecessary string copy. In the 2nd alternative the caller code fails
to compile. If instead the caller is written like this:
  const auto& name = foo.getName();
  if (name.empty()) { ...
Then all variants compile fine and don't introduce unnecessary copies.

Usually the caller code doesn't care exactly what type a function returns. It
just care about what it can do with the returned object. E.g. in this example
it wants to test whether the returned string-like object is empty. Using auto
helps to make the code type-agnostic and avoids introducing unnecessary
conversions. The openMSX code did contain a few cases that made unnecessary
string copies.

Another example where type-agnostic code can improve correctness is the
following:
  auto size = file.getSize();
  if (0x10000 <= size && size < 0x20000) { ...
Initially in openMSX our get-file-size method returned a value of type
'unsigned' (32-bit), later we changed that to return a 64-bit value (on 64-bit
platforms). If the caller would use the (old) type 'unsigned' instead of 'auto'
then after the get-file-size refactoring, the caller code has a bug (and the
compiler won't complain about the code). If instead auto is used then the same
caller code is correct before and after the refactoring.