Use initializer lists (new in c++11)
Visual studio 2013 should now support these. Let's see if it true and fully
compatible with gcc ;-)

Mostly use initializer lists to construct collections in a single statment
instead of repeated calls to e.g. push_back(). Because of this the code may
also be more efficient, but none of these were in a critical part of the code.