Store old value of machine specific update events in MSXCliComm instead of GlobalCliComm
To avoid sending too much data over clicomm connections, we remember the last
value of an update event and only send the event if the new value is different.

Before this commit this was handled in one place (GlobalCliComm) for both
global and machine specific update events. This commit duplicates this filter
mechanism in both GlobalCliComm (now only global) and MSXCliComm (only machine
specific).

The advantage is that when a machine is deleted, the data about the machine
specific events now also disappears. Before this data was kept forever (about
5kb per machine). This is a memory leak, but not one detectable by tools like
valgrind (because the memory is properly freed when openmsx exits).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8239 7512d703-071c-0410-b5c2-d2bdd895b7d1