fixed assert on "unset renderer ; loadstate"
This assert was triggered when changing renderer and very soon after do a
loadstate (Quibus first hit this problem in catapult).

The problem is that, for reasons I won't explain here (see comments in the
code), switching renderer cannot be done immediately. Instead an event is
pushed on the event queue and switching renderer is only actually done when
this event is processed. This causes problems in the 'loadstate' command
because the renderer settings is already changed, while the actual renderer is
not yet changed.

I extended the code to deal with this situation. As a bonus the V99x8 and V9990
code is now also more symmetrical.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8278 7512d703-071c-0410-b5c2-d2bdd895b7d1