fixed assert in Scheduler on loadstate
MSXMixer (indirectly) registers syncpoints in the constructor of
MSXMotherBoard. This is before the state of the Scheduler is restored from the
savestate. So it is possible (even very likely) that the MSXMixer syncpoints
are now before the restored 'scheduleTime' of the Scheduler. This will trigger
an assert in the Scheduler when the MSX starts running.

However by luck these syncpoints got removed before the machine actually starts
running (they are removed because the MSXMixer is unmuted before the machine
starts, and because of implementation details of unmute). But in case of a
savestate of a turbor in the 'HW-muted' state, the mixer is not unmuted and the
assert does get triggered.

Solved by always rescheduling the MSXMixer syncpoints after the Scheduler is
de-serialized.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8279 7512d703-071c-0410-b5c2-d2bdd895b7d1