fixed 'hanging OSD leds' when switching between savestates:
For example when switching to a non-turbor savestate while the 'turbo' LED was
on, this turbo LED remained active. Also the LEDs became completely
unresponsive after the last msx machine was deleted (delete_machine command).

It seems that using the Tcl 'trace variable' command on settings is trickier
than I originally thought, especially if the settings can (temporarily) go away.
Maybe in the future we should create a 'after setting_changed' command to make
this easier.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8281 7512d703-071c-0410-b5c2-d2bdd895b7d1