improved earlier fix (revision 7746):
Settings were looked up by name. Though it's possible that when a setting
disappears, by the time the callback gets triggered, a new setting with the
same name already appeared. Now the lookup is done using a unique (always
increasing) number.

I didn't find a situation where the old code went wrong. Maybe ATM nothing goes
wrong when the callback acts on the new Setting. But it anyway feels better to
fix this.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8285 7512d703-071c-0410-b5c2-d2bdd895b7d1