added "gz-base64" blob encoding
First the blob is gzipped, and then base64 encoded. Compared to the previous
encoding (hex), this results in smaller savestates.

The compressed savestates are somewhat smaller, but the uncompressed savestates
are a lot smaller. This is important on embedded systems with not too much RAM
(like GP2X) because the uncompressed savestate is stored in memory at some
point in time.

As a side-effect, loading/saving savestates is also a little faster now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8290 7512d703-071c-0410-b5c2-d2bdd895b7d1