fixed reading of gzipped xml files
Libxml can optionally be configured to support gzipped xml files. I
(unintentionally) relied on this feature. For example the standard installed
libxml on the GP2X is build without this feature, but the version installed on
my Debian system does have it.

Fixed by adding code that uncompresses the file before passing it to libxml.
Though when libxml _does_ support gz, we still use that feature (in that case
we still pass the compressed file directly) because that's likely more
efficient. Luckily there's an easy way to test how libxml was configured.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8291 7512d703-071c-0410-b5c2-d2bdd895b7d1