templatize VDPCmdEngine command objects on LogOp
The command objects were already templatized on the screen mode. Now the
objects that implement logical operations are also templatized on this
operation. This results in faster (more specialized) code.

Before the logical operation was called via a virtual function. This is slower,
not because the virtual call itself is so much slower, but more because it
cannot be inlined and thus prevents further compiler optimizations.

A disadvantage of this approach is that it generates a lot more code
(specialized code for each combination of command/screen-mode/logop). I
measured an increase in code size of about 280kB (measured after the cleanups I
will commit next). However a lot of this code is rarely executed and the code
paths that are frequently executed are shorter, so the total working set might
even be smaller.

In the current code there is still a memory leak. It will be fixed in the next
commit (together with some more cleanups). I didn't want to commit both patches
at once to make the functional change (= this patch) easier to understand.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8309 7512d703-071c-0410-b5c2-d2bdd895b7d1