implemented 'blink' fetaure in bitmap modes
This is related to bug:
  [ 2212686 ] Blink not working on graphic modes

The testcase we got from NYYRIKKI works now. I didn't do tests on a real
machine to figure out a model for this behaviour. Instead I copied the
implementation from MESS. And this seems a very logical way for how this
feature can be implemented in the real HW (matching very closly the odd/even
page flipping).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8338 7512d703-071c-0410-b5c2-d2bdd895b7d1