fixed flags in R800 emulation, rexall passes now
I updated the CRC constants of the zexall test to match the values meausured on
a real R800. Then I fixed the R800 emulation to match those results. Or IOW if
you run the unmodified zexall test, it will now fail with the same numerical
CRC errors as on the real HW.

The biggest difference between the flags on Z80 and R800 is that practically
all R800 instructions leave the (undocumented) X and Y flags unchanged, while
most Z80 instructions do put some result in them. But there are other
differences, for example in the CCF, BIT and SLI instructions and also in some
invalid instructions, see source code for details.

The zexall test is not complete. Not for Z80, but certainly not for R800. So
there might come more changes like this in the future.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8341 7512d703-071c-0410-b5c2-d2bdd895b7d1