removed sectorcache code from DirAsDSK
Code was already effectivly disabled (in the code was a comment ".. still need
to implement this ..". It was like this for over a year, and nobody seemed to
miss it. So now I removed that code.

The sectorcache had two goals:
 - Before dirasdsk had 'full synchronization' with the host filesystem, the
   sectorcache was useful to keep changes (done on MSX) that extend the size
   of a file.
 - Some (cracked) games load the game data from files, but still use some fixed
   sectors for the in-game savestates.

The first case is no longer valid and the second is very rare. I discussed this
with David and Alex and we agreed to remove the sectorcache.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8346 7512d703-071c-0410-b5c2-d2bdd895b7d1