pass EmuTime by-value on 64-bit CPUs
On 64-bit CPUs the EmuTime struct fits in one register, so in terms of
registers it's equally cheap to pass-by-value or pass-by-reference. However
actually using the value is cheaper in the former case (you save the
dereference operation). I also measured that the resulting code is slightly
smaller and slightly faster.

On 32-bit CPUs we still use pass-by-reference-to-const.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8352 7512d703-071c-0410-b5c2-d2bdd895b7d1