fixed 'slot-a already in use' error on loadstate:
You could, for example, trigger it like this:
  - start openmsx with  'openmsx <rom1> <rom2>'
  - in openmsx console: 'carta >rom3>'
  - savestate / loadstate

The problem was that the savestate still contained 'any' slot specifications
and the order in which the cartridges are instantiated at loadstate can be so
that the 'any' slot and the fixed 'slot-a' collide.

Fixed by always saving the actual slot in a savestate.

Note that savestates create before this fix can still show this problem.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8374 7512d703-071c-0410-b5c2-d2bdd895b7d1