fixed off-by-one error in VDPCmdEngine:
In graphic4 mode (screen 5), executing a high-speed VDP command with logical
operation number not equal to zero (this number is not used for high-speed
commands), triggered a crash, because part of an array was not correctly
initialized. Simply booting a Panasonic FS-A1F machine (the builtin firmware)
triggered this. Fixed now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8409 7512d703-071c-0410-b5c2-d2bdd895b7d1