Fixed "trainers don't work after loadstate" problem
After a machine switch (part of loadstate), the 'after time' mechanism that
drives the trainer gets deactivated. But the trainer bookkeeping in Tcl
variables still indicates there is an active trainer. This of course confuses
the trainer script with as effect that you cannot reactivate the trainer
(because the script thinks it is already active).

Solved by letting the script also listen to the 'machine_switch' event and
clear the bookkeeping when that event occurs.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8620 7512d703-071c-0410-b5c2-d2bdd895b7d1