Fixed crash: invoke signalEvent() on deleted object
Could be triggerd like this:
 start openmsx with the ide extension in slot-a:
    openmsx -ext ide
 replace the ide extension with something else via the console:
    carta aleste.rom

The problem was that when inside the loop of MSXEventDistributor::distributeEvent()
some MSXEventListener unregisters itself from the event distributor,
there is still a chance that the currently being distributed event gets
delivered to the just unregistered object. (And in this case that object
was already deleted.)

Thanks to Vampier for reporting and to mfeingol for finding the
cause of this problem.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9321 7512d703-071c-0410-b5c2-d2bdd895b7d1