fixed [1726234] Diskmanipulator: assert when disk image is corrupt
Actually there are two fixes, the one described in the bug report and also
an assert when using a diskimage of size zero (for some strange reason there
was such a disk image in my collection).

This fix required to move the method Disk::isReady() to the base class
SectorAccessibleDisk. I also renamed it to isDummyDisk(). And similarly
I renamed DiskDrive::isReady() to isDiskInserted().

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9383 7512d703-071c-0410-b5c2-d2bdd895b7d1