extended load_icons script with 'pause', 'mute' and 'breaked'
I know, the icons I drew are ugly and really need to be replaced. Feel free to
draw some better ones.

The current icons are also too big. They overlap when placed left or right and
go too low when placed at the bottom (only top view is right at the moment).
Also the position needs to be fixed in the 'fancy' skin. But all this can only
be done once we have new properly sized icons.

Apart from all this it does work ;-)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9399 7512d703-071c-0410-b5c2-d2bdd895b7d1