Refuse to load rom images with improper filesize
E.g. for a rom mappertype that works with banksize of 8kB, the uncompressed
filesize of the rom image must be an integer multiple of 8kB. This is an
extra safety mechanism to detect a wrongly selected mappertype.

It also fixes bug
  [2710707] The last page of a megarom is not loaded on hacked roms
Instead of silently ignoring the extended part of the (invalid) rom image,
the user is now informed that something is wrong with the rom image.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9423 7512d703-071c-0410-b5c2-d2bdd895b7d1