Fixed mute on 'debug break'
In revision 9372 we changed the 'debug break' behaviour to globally block
openmsx as opposed to only the currently active MSX machine. Though the
mute-on-break functionality was implemented at the machine level.

Fixed now by moving this functionality to the global level. As a bonus the
new code also handles mute-on-pause, so that functionality could be removed
from the Mixer.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9443 7512d703-071c-0410-b5c2-d2bdd895b7d1