Fixed [1804669] Crashes possible when wrong ROM is used with a mapper
Part of the fix was done in revision 9423 (refuse to load rom images with
improper filesize). Only two mappers needed additional checks.

The MSX-AUDIO mapper type that was mentioned in the bug report has since been
removed. I checked the old code for the cause of the crash and audited all
current mapper types for the same kind of bug.

While modifying the RomFSA1FM code I noticed that the code only uses the first
1MB of the ROM image, but the dumped image I have (with matching SHA1SUM) is
2MB in size. And indeed the 2nd half of that dump is almost identical to the
first half (see comments in RomFSA1FM for why it's not completely identical).
So I created such a 1MB dump and adjusted the SHA1SUMS accordingly.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9444 7512d703-071c-0410-b5c2-d2bdd895b7d1