Added support for taking screenshots without OSD elements
Restructured OutputSurface and child classes to make it possible to render to
an off-screen surface, so that it becomes possible to re-render the current
frame without the OSD layers.

The default is now to create a screenshot without OSD elements, because we
now (as default) have a pause icon that won't fade away. To create screenshots
with OSD included, pass the -with-osd option to the screenshot command.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9511 7512d703-071c-0410-b5c2-d2bdd895b7d1