Simplified power-off/reset handling in MSXMotherBoard
In the past (already several years ago) it was needed to handle power-off and
reset in MSXMotherBoard in two phases: a first phase that only sets a flag that
the action needs to happen and a second phase that actually performs the
action. The first phase could run while the Z80 was still active somewhere in
the callstack. The second phase was guaranteed to run without 'active' MSX
peripherals.

Several years ago we already changed the handling of commands so that they all
run without MSX-device code active on the stack. So there is no reason anymore
to split reset and power-off in two phases. This allows to simplify the code
quite a bit.

It also removes a bug. The old code didn't work correctly when the
MSXMotherBoard was in 'debug-break' state. This is bug
  [2776810] openMSX debug_break + turn-off/on bug

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9653 7512d703-071c-0410-b5c2-d2bdd895b7d1