Fixed bug with blur=1
With renderer=SDL, blur=1, scale_factor=3 and in a 256 pixel wide MSX screen
mode, the scaling calculation went wrong: it showed a black screen.

Problem was an overflow in 16-bit fixed-point calculation. We use 0.16 fixed
point numbers, but that format can't represent '1.0'. The same would happen
with blur=0, but there was already a check for that. Fixed by generalizing this
existing check so that it catches all cases were this overflow would occur.
(Actually the check was there to improve performance with blur=0, but it's
also needed for correctness).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9751 7512d703-071c-0410-b5c2-d2bdd895b7d1