Fixed debuggable size of MSXRam.
All debuggables associated with MSXRam had a size of 64kB, even if in
hardwareconfig.xml they were given a smaller size.

For example

  <RAM id="Main RAM">
    <mem base="0xC000" size="0x4000"/>
  </RAM>

The RAM tag has default values for some attributes, so the above is actually
the same as

  <RAM id="Main RAM" base="0x0000" size="0x10000">
    <mem base="0xC000" size="0x4000"/>
  </RAM>

The base/size pair on <RAM> indicates the actual size of the RAM, the base/size
pair in <mem> indicates were in memory this should be visible (this is still a
leftover from before we has the <mem> tag).

So actually the debuggable had the proper size, but the configuration format is
not very intuitive. None of our current machine configs put base/size
attributes on the <RAM> tag. So I've removed support for those tags. Instead
that info is derived from the inner <mem> tag(s). With this change, the example
above creates a debuggable with size 16kB.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9804 7512d703-071c-0410-b5c2-d2bdd895b7d1