Fixed IE0 related VDP bug:
When there is a pending VBLANK IRQ (bit 7 in S#00 is set) and you switch IE0
(bit 5 in R#1) on, there is immediately a IRQ send to the Z80. Before this fix
openMSX would only generate one on the next VBLANK.

This fixes two long-standing bugs:
 - [1849488] Game "Adonis" plays sound way too slowly
 - [ 739150] Zanac: title corrupted

Thanks a lot to 'n_n' for providing a z80 trace of a Adonis run in meisei.
Comparing this trace with a openMSX trace allowed to pin-point this bug.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9983 7512d703-071c-0410-b5c2-d2bdd895b7d1