better implement difference between power-on and reset
For example clear the (V)RAM content after a power-down / power-on.
This is bug:
    [2776812] set power off don´t clean memory contents

An alternative implementation would be to remove all powerUp() and powerDown()
methods and instead delete/recreate the MSX machine. Though that would also
loose the state of the Pluggables (which are currently connected) and the
inserted disks/tapes.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10107 7512d703-071c-0410-b5c2-d2bdd895b7d1