Refactored YM2413 cores to allow easier testing
The YM2413 cores are now almost completely independent from the rest of openMSX
(the only exception is serialization). This makes it easier to test the core in
isolation. It also makes it easier to reuse the code in different emulators.

There is a start of a unittest (YM2413Test.cc) but ATM this code is not yet
executed, or even compiled by default.

See YM2413Core.hh for lots more details about the new interface.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10124 7512d703-071c-0410-b5c2-d2bdd895b7d1