optimized serialize to memory
Use new OutputBuffer/InputBuffer class to create/load in-memory savestates
instead of std::vector.

These new classes are much more specialized for the job compared to std::vector
(see comments in the code for details). So this results in noticeably faster
in-memory snapshots (will be important for the planned 'reverse' feature).
The resulting code is also about 60kB smaller.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10138 7512d703-071c-0410-b5c2-d2bdd895b7d1