made 'serialize without ID' the default
While serializing a struct (a struct 'by value' or a pointer to a struct), we
give this struct an ID. This way when we later serialize (another) pointer to
this struct we can simply refer to the previous ID.

This mechanism works very well, but in openMSX it's only needed a couple of
times. So we changed the default from include the ID (and exclude on demand) to
exclude the ID (and include on demand).

This results in smaller savestates. Loading/saving is also faster. And the
openMSX executable is about 80kB smaller.

There is no problem with loading older savestates. The extra (unused) IDs are
simply ignored. Nor is there a problem with loading the new savestates on an
older openMSX.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10139 7512d703-071c-0410-b5c2-d2bdd895b7d1