pass std::string by pointer
Changed function parameter from
  std::string type = ""
into
  std::string* type = NULL

It turns out that constructing an empty string object (with associated unwind
code for the case that constructing the empty string throws an exception) takes
quite some memory. The executable becomes 120kB smaller if we pass the string by
pointer and let a NULL pointer represent an empty string.

This function is called A LOT of times that's why it makes such a noticable
difference. It wouldn't recommend this transformation for normal code.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10140 7512d703-071c-0410-b5c2-d2bdd895b7d1