use __builtin_constant_p to simplify calls to OutputBuffer::insert()
'__builtin_constant_p' is a gcc extension, it allows to test whether a certain
expression is a compile-time constant or not. We use it to detect calls to
OutputBuffer::insert() with the num parameter equal to 1, 2, 4, or 8. In such
cases we call a specialized version of the routine, in all other cases we call the
generic version.

The benefit here does not come from the specialized code itself, but from the
fact that the specialized function has one parameter less. This means the
caller has to setup one parameter less and this saves a couple of bytes per
call. Though there are a lot of callers and _by_far_ the majority has the num
parameter equal to 1, 2, 4 or 8. The executable became about 35kB smaller with
this patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10143 7512d703-071c-0410-b5c2-d2bdd895b7d1