pass 'const char*' instead of 'const std::string&' to beginTag() and endTag()
The methods beginTag() and endtag() intrenally need a std::string. Still we
changed them to instead take a 'const char*' parameter.

All callers of these two methods pass a string literal (thus a const char*) as
parameter. If the method takes a std::string then all callers need to construct
a string object. If the method takes a const char*, then only the method needs
to convert the const char* into a std::string.

This saves about 270kB on the executable size!

An alternative solution would be to overload beginTag() with both a std::string
and a const char* version and then let the const char* version simply call the
std::string version. This trick might be useful in other places as well (e.g.
when registering a setting or command we also mostly pass a string literal to a
method that expects a std::string).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10144 7512d703-071c-0410-b5c2-d2bdd895b7d1