added minilzo
See http://www.oberhumer.com/opensource/lzo/

LZO is a very fast (de)compression routine. Though compression ratio is not as
good as e.g. GZIP. The source code is small enough to be included in the
openMSX source tree (and including the source code is also the recommended way
according to the LZO documentation). Except for renaming .c -> .cc and .h ->
.hh, the sources are unchanged compared to the original distribution.

Code is not yet used in this commit. It will be in the next commit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10191 7512d703-071c-0410-b5c2-d2bdd895b7d1