overload several construtors that take string params
SimpleCommand, Command and XXXSetting constructors all take one or more
std::string parameters. Very often these are called with string literals,
so each caller must convert that literal (const char*) into a string itself.

In this patch those constructors are overloaded to also take const char*
parameters. (It's overloaded and not simply replaced because in some cases
we do pass actual std::string parameters). The implementation of the
constructors is moved to a common helper routine (if the body was not empty).
(The next c++0x standard will allow for one constructor to call another
constructor, like in Java. This would allow to simplify this patch).

This saves about 35kB (linux x86_64, devel flavour).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10226 7512d703-071c-0410-b5c2-d2bdd895b7d1