fixed bug [2823673] moonsound 4 operator FM fail
Bug was introduced in revision 8151 (2008/08/08). It was a (micro) optimization
to speed up OPL4 FM. I reverted it for now.

TODO need to figure out why this optimization was not correct and fix it (but
first I need to get some sleep). If it's worth it at all, it really is only a
very minor optimization. Though at the time I hoped it could help detangle the
YMF262 code mess and allow further simplifications and/or speedups.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10233 7512d703-071c-0410-b5c2-d2bdd895b7d1