break cyclic dependency between GlobalCliComm and GlobalCommandController
It was possible to crash openmsx by making your settings.xml file read-only
and exiting openmsx (with save-settings-on-exit enabled): Destructing
GlobalCommandController (indirectly) saves the settings. When settings.xml
could not be written (e.g. because it's read-only) we will print a warning.
But (before this patch) the GlobalCliComm object, which is responsible for
distributing the warning message, was already deleted. Fixed by swapping
the destruction order of cliComm and commandController. Though to make
this possible, the cyclic dependency between the two had to be broken:

GlobalCommandController depended on GlobalCliComm:
  a) For message printing.
  b) Indirectly other classes request a reference to cliComm via
     commandController.
GlobalCliComm depended on GlobalCommandController:
  c) Because it owns the 'update' command.
  d) Because it created StdioConnection and PipeConnection objects which need
     commandController (to execute the commands send over clicomm connections).

Dependency a) is a valid dependency. b) should ideally disappear in the future.

c) is removed by moving the update command to GlobalCommandController (and
adding a GlobalCliComm::getUpdateStrs() method to query the (now shared) list
of 'update names').

d) is removed by moving initial clicomm connection creation (stdio or pipe) to
CommandLineParser. (Setting clicomm in XML mode is now done via a new
GlobalCliComm::setXMLOutput() method).

So now cliComm doesn't need commandController anymore. But CliConnection
objects,  which are owned by GlobalCliComm, still use commandController. So
between the time commandController and cliComm are destructed, there are still
CliConnection objects alive which refer to the (deleted) commandController
object. CliConnection uses commandController (only) to execute commands
received over the connnection. These commands are received in a helper thread
and thus it can potentially happen at this moment in time. But still this is
not a problem because the actual command execution is triggered via events
(helper thread triggers event in main thread) and events are no longer
processed at this point in the shutdown sequence.

In addition to fixing the above problem, this patch also allows to get rid of
the ugly getCliCommIfAvailable() and setCliComm() methods.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10236 7512d703-071c-0410-b5c2-d2bdd895b7d1