Also support laserdisc for scale_factor 1 and 3
Extended the existing TransparentScaler class to also handle interlaced V99x8
modes (not yet required, but possibly in the future). Resolutions that are
specific to Gfx9000 are not yet(?) supported, but they will print a warning
when they're used. Also renamed this class to Transparent2xScaler.

Implemented Transparent1xScaler and Transparent3xScaler with support for the
same modes as above.

Note that the Transparent Scaler is not always selected when you switch
scale_factor (you have to reset or even restart openMSX). I'll have to
investigate and fix this bug.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10254 7512d703-071c-0410-b5c2-d2bdd895b7d1