fixed inconsistency in VDP superimposing state
The VDP internally calculates the 'are we currently superimposing' state at the
beginning of a frame and send an update if this state changed. However if this
status was queried from external code, the result was based on current register
values. Fixed by making the query also return the per-frame value.

The change in SDLRasterizer::setSuperimposing() is not a bug fix, but the old
code was very confusing to me. Now it's implemented similar to the other color0
related update methods (e.g. setTransparency()).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10259 7512d703-071c-0410-b5c2-d2bdd895b7d1