fixed [2828096] changing scaler when paused causes glitches
The problem was that the hq and hqlite openGL scalers required extra data
(edgeBuffer) that should be uploaded as a texture. But this was only done
when new rawframe data became available. So not when in pause mode.

Fixed by re-uploading framedata on scaler switch. See also comments in
src/video/GLPostProcessor.cc (method paint() line 155).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10290 7512d703-071c-0410-b5c2-d2bdd895b7d1