bugfix with multiple ranged io watchpoints
It was possible to crash openMSX when there were multiple io watchpoints that
were associated with a range of IO ports and where that range overlapped. For
example like this:
  debug set_watchpoint read_io {0x00 0x30}
  debug set_watchpoint read_io {0x10 0x40}
  debug remove_watchpoint wp#1
The root of the problem was that the internal WatchPointIODevice only had one
pointer to the real MSXDevice, even it it spanned a whole IO-port range.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10319 7512d703-071c-0410-b5c2-d2bdd895b7d1