Fixed behaviour of YMF278 'NEW2' bit
Writes to the wave part should be ignored when NEW2=0 (see comments in source
code for details). Before this patch this functionality was missing.

There was also a bug in the code for reading the status register. When the NEW2
bit is set (only once after reset) bits 2,1 in the status register are set to
0,1 (again see comments in source code for more details). There was a bug that
coupled this behavior to the 'NEW' bit instead of 'NEW2'.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10325 7512d703-071c-0410-b5c2-d2bdd895b7d1