Added OSD mouse support
Made it possible to query the mouse coordinates in the openMSX OSD system.
On every widget there is now a (read-only) -mousecoord property. This
property returns the coordinates of the mouse relative to the bounding box
of the widget. So when the mouse is over the widget, the coordinates will
be inside (0,0)-(1,1).

This makes it possible to
 - Query the absolute mouse coordinates (query the root widget).
 - Query MSX mouse coordinates (using the MSX OSD widget from the existing
   osd_widgets.tcl script).
 - Easily test if the mouse is over a specific widget.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10348 7512d703-071c-0410-b5c2-d2bdd895b7d1