Fixed link error with vc++
Before this patch we had a link error with vc++, but not with gcc. Turns
out vc++ was wrong. See this thread for more info:

   http://www.codeguru.com/forum/showthread.php?t=430949

So even though the default vc++ behaviour might be easier, it's not
standard compliant, so a pain to work with in portable code :(

Thanks to mfeingold to investiagte this.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10360 7512d703-071c-0410-b5c2-d2bdd895b7d1