Fixed crash with memory watchpoints
Before this patch, if you removed a watchpoint, the memory cache administration
was off, so that it still thought there was a watchpoint on that region (was
fixed on the next mem watchpoint add/remove). This bug was there for years, but
the only effect was that emulation was slightly slower after removing a
watchpoint (speed went not back to the level without any watchpoints). Though
recently with the additional of the watchpoint Tcl variables, we added some
asserts that check that if a watchpoint is triggered in a certain region, there
really must be at least one watchpoint defined. So this triggers on this old
bug.

Administartion is fixed now. Thanks to BiFiMSX for reporting this problem.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10363 7512d703-071c-0410-b5c2-d2bdd895b7d1